chore: move saga navigation utils inside correct folder
Update deprecated with jsDoc since it's not possible to use navigation inside the saga
