--- ts/features/bonus/cgn/navigation/actions.ts
@@ -1,106 +1,0 @@
-import { CommonActions } from "@react-navigation/native";
-import NavigationService from "../../../../navigation/NavigationService";
-import { CgnMerchantDetailScreenNavigationParams } from "../screens/merchants/CgnMerchantDetailScreen";
-import { CgnMerchantLandingWebviewNavigationParams } from "../screens/merchants/CgnMerchantLandingWebview";
-import CGN_ROUTES from "./routes";
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnActivationInformationTos = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.INFORMATION_TOS
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnActivationLoading = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.LOADING
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToEycaActivationLoading = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.EYCA.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.EYCA.ACTIVATION.LOADING
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnAlreadyActive = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.EXISTS
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnActivationPending = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.PENDING
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnActivationTimeout = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.TIMEOUT
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnActivationIneligible = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.INELIGIBLE
-    })
-  );
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnActivationCompleted = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.ACTIVATION.MAIN, {
-      screen: CGN_ROUTES.ACTIVATION.COMPLETED
-    })
-  );
-
-// Details
-
-/**
- * @deprecated Do not use this method when you have access to a navigation prop or useNavigation since it will behave differently,
- * and many helper methods specific to screens won't be available.
- */
-export const navigateToCgnDetails = () =>
-  NavigationService.dispatchNavigationAction(
-    CommonActions.navigate(CGN_ROUTES.DETAILS.MAIN, {
-      screen: CGN_ROUTES.DETAILS.DETAILS
-    })
-  );
