+++ ts/features/bonus/cgn/saga/orchestration/eyca/eycaActivationSaga.ts
@@ -7,7 +7,7 @@ import { BackendCGN } from "../../../api/backendCgn";
 import {
   navigateToCgnDetails,
   navigateToEycaActivationLoading
-} from "../../../navigation/actions";
+} from "../navigation/actions";
 import {
   cgnEycaActivation,
   cgnEycaActivationCancel
